requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.7
pydantic-settings>=2.3
passlib[bcrypt]>=1.7.4
PyJWT>=2.8
python-multipart>=0.0.9
email-validator>=2.0.0
bcrypt==4.0.1
chromadb>=0.5.5
sentence-transformers>=2.7.0
numpy>=1.26
pytest>=7.4